Implemented generic query builder to bridge differences between DBs.
The query builder supports generic type names, but they are not yet used in the
implementation.

Also builds on Windows (64-bit), but the necessary build files are not yet
committed, as I am not yet sure what to include in the repo.

+++ Core/Io/Url.cpp
@@ -502,7 +502,7 @@ namespace storm {
 	Url *executableFileUrl(Engine &e) {
 		std::vector<wchar_t> buffer(MAX_PATH + 1, 0);
 		do {
-			GetModuleFileName(NULL, &buffer[0], buffer.size());
+			GetModuleFileName(NULL, &buffer[0], DWORD(buffer.size()));
 			if (GetLastError() == ERROR_INSUFFICIENT_BUFFER) {
 				buffer.resize(buffer.size() * 2);
 				continue;
